Open up the deleteAll endpoint to all models by default
