--- mdm-plugin-referencedata/grails-app/utils/uk/ac/ox/softeng/maurodatamapper/referencedata/rest/transport/DeleteAllParams.groovy
@@ -1,38 +1,0 @@
-/*
- * Copyright 2020-2022 University of Oxford and Health and Social Care Information Centre, also known as NHS Digital
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-package uk.ac.ox.softeng.maurodatamapper.referencedata.rest.transport
-
-import grails.validation.Validateable
-
-/**
- * @since 04/09/2018
- */
-class DeleteAllParams implements Validateable {
-
-    List<String> ids
-    Boolean permanent
-
-    DeleteAllParams() {
-        permanent = true
-    }
-
-    static constraints = {
-        ids minSize: 1
-        ids nullable: false
-    }
-}
